+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # CHANGELOG
 
+**v0.1.3:**
+- Introduced new services as [social icons](examples/ConfirmEmailExample/Program.cs):
+  - `Service.Phone` (converts into a link with `tel:` prefix)
+  - `Service.Facebook`
+  - `Service.YouTube`
+  - `Service.Instagram`
+  - `Service.LinkedIn`
+  - `Service.Slack`
+  - `Service.Discord`
+  - `Service.TikTok`
+  - `Service.Snapchat`
+  - `Service.Threads`
+  - `Service.Telegram`
+
 **v0.1.2:**
 - Introduced `CHANGELOG.md`
 - Added `icon.png` to `PackageIcon`